fix: move test files to tests dir
